Strip ./ prefix from commands on Windows to avoid shell error.
This change fixes "'.' is not recognized as an internal or external
 command" errors that occur when trying to build with conan from
inside a msys2/mingw64 shell window. Even though conan is being
launched from a mingw64 bash shell, it invokes the Windows shell
when running commands for recipes that do not have win_bash set
and cmd.exe does not handle commands that start with "./".